Updated simulation values
Update the simulation parameters and interpolation range in order to get closer to the numerical result by GR. I have al
so updated the example script, plots and added a short description of what I have been doing. This must still be complet
ed. Furthermore, the publication scripts must be updated and a summary must be written.